+++ custom_model_runner/datarobot_drum/drum/typeschema_validation.py
@@ -4,6 +4,7 @@ All rights reserved.
 This is proprietary source code of DataRobot, Inc. and its affiliates.
 Released under the terms of DataRobot Tool and Utility Agreement.
 """
+import numbers
 import os
 import sys
 from abc import ABC, abstractmethod
@@ -270,12 +271,15 @@ class DataTypes(BaseValidator):
     @staticmethod
     def is_img(x: pd.Series) -> bool:
         def convert(data):
+            if isinstance(data, numbers.Number) and np.isnan(data):
+                return np.nan
             return Image.open(BytesIO(base64.b64decode(data)))
 
         try:
             x.apply(convert)
             return True
-        except:
+        except Exception as e:
+            print(e)
             return False
 
     @staticmethod